requirements.txt
appdirs==1.4.3
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
cPyparsing==2.3.2.1.0.0
cycler==0.10.0
Cython==0.29.6
decorator==4.0.10
enum34==1.1.6
freetype-py==1.0.2
future==0.18.2
ghalton==0.6.2
ipython==5.1.0
ipython-genutils==0.1.0
kiwisolver==1.1.0
lazy==1.4
Mako==1.0.4
MarkupSafe==0.23
matplotlib==3.1.2
moderngl==5.5.4
moderngl-window==2.0.5
ModernGL.ext.obj==1.0.0
multipledispatch==0.6.0
numpy==1.17.4
OpenEXR==1.3.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==6.2.1
pkg-resources==0.0.0
pocl-binary-distribution==1.3.post1
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pycparser==2.14
pyglet==1.4.9
Pygments==2.1.3
pyopencl==2019.1.2
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.4
pyparsing==2.4.5
PyQt5==5.13.2
PyQt5-sip==12.7.0
PyQtWebEngine==5.13.2
pyrr==0.10.3
python-dateutil==2.8.1
pytools==2019.1.1
Shapely==1.6.4.post2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
